+++ client/src/pages/add.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { Form, FormDescription, FormItem } from "@/components/ui/form";
-import { WearableAddCard } from "@/components/ui/wearable-add-card";
-import { WearableFileInputButton } from "@/components/ui/wearable-file-input-button";
+import { WearableAddCard } from "@/components/wearable-add-card";
+import { WearableFileInputButton } from "@/components/wearable-file-input-button";
 import { useCreateWearables } from "@/hooks/api";
 import { useToast } from "@/hooks/use-toast";
 import { zodResolver } from "@hookform/resolvers/zod";
